# Speed-up math.dist() by 30%